% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalence.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Predicting Prevalence from a Bayesian Mixed or Fixed Effect Logistic
Regression with Presence/Absence Tests on Pooled Samples}
\usage{
getPrevalence(model, data = model$data)
}
\arguments{
\item{model}{An object of class \code{brms} with the regression outputs from
a call to \code{PoolRegBayes}}

\item{data}{The data for which prevalence needs to be estimated/predicted. If
not provided, defaults to using the data used to train the model (i.e.
returns the fitted values of the prevalence)}
}
\value{
A \code{list} with at least one field \code{PopulationEffects} and an
  additional field for every random/group effect variable. The field
  \code{PopulationEffects} contains a \code{data.frame} with the prevalence
  estimated based only the fixed/population effects. When the intercept is
  the only fixed/popultion effect, this is just the population mean (possibly
  adjusted for random/group effects). When there are also
}
\description{
This function works somewhat like a \code{predict} or \code{fitted} generic
function returning the model predicted prevalence for a given set of data;
however, as the quantity of interest (prevalence) is neither on the response
or link scale we do not use either of these generic functions. Further, when
the model accounts for the heirarchical structure of the sampling frame (e.g.
Region/Village/Site), it is typical to want to know the predicted values at
each level of sampling (e.g. Prevalence at each region, village or site) so
these are calculated automatically.
}
