% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogitRegression.R
\name{PooledLogitRegression}
\alias{PooledLogitRegression}
\title{Logistic regression with presence/absence tests on pooled samples}
\usage{
PooledLogitRegression(data, PoolSize, formula, alpha = 0.05, verbose = F)
}
\arguments{
\item{data}{A \code{data.frame} with one row for each pooled sampled and columns for
the size of the pool (i.e. the number of specimens / isolates / insects
pooled to make that particular pool), the result of the test of the pool
and any number of columns to be used as the dependent variables in the
logistic regression}

\item{PoolSize}{A \code{string} giving the name of the column with number of specimens/isolates/insects in each pool}

\item{formula}{A \code{formula} of the kind used to define models in \code{glm}. The left-hand
side should be the name of column with the result of each test on each pooled
sample. The result must be stored with 1 indicating a positive test result and
0 indicating a negative test result.}

\item{alpha}{The confidence level to be used for the confidence and credible intervals.
Defaults to 0.5\% (i.e. 95\% intervals)}

\item{verbose}{Logical indicating whether to print progress to screen. Defaults to false
(no printing to screen)}
}
\value{
TBC
}
\description{
Logistic regression with presence/absence tests on pooled samples
}
\examples{

#Same synthetic data with four locations (A-D) and three sampling years (0-2)
#Locations A-C have different baseline prevalences.
#Prevalence at D = Prevalence at A.
#All locations have the same expoential growth rate, -0.5, i.e. falling prevalence

N <- 5000
DataTrend <- data.frame(Place = sample(c("A","B","C","D"),N, replace = T),
                        Year = sample(c(0:2), N, replace = T),
                        NumInPool = sample(25:30, N, replace = T)
                        )
BasePrev <- c(A = 0.1, B = 0.05, C = 0.02, D = 0.1) #The baseline prevalences at each location
GrowthRate = -0.5 #The (negative) growth rate
#Calculate the true prevalence at each time and place
DataTrend$TruePrev <- with(DataTrend,
                          BasePrev[Place] * exp(GrowthRate*(Year-min(Year))))
#Simulate some pooled samples from each location
DataTrend$Result <- with(DataTrend,
                         runif(N) < 1-(1-TruePrev)^NumInPool)

#Calculate the prevalence at each location and timepoint
Prevs <- PoolPrev(DataTrend,Result,NumInPool,Place,Year)
Prevs

#Perform logistic regression with explanatory variables Place and Year
Reg <- PooledLogitRegression(DataTrend,"NumInPool",Result ~ Year + Place)
Reg$OR
}
