% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledRegressionAlt.R
\name{PooledRegressionAlt}
\alias{PooledRegressionAlt}
\title{Logistic regression with presence/absence tests on pooled samples}
\usage{
PooledRegressionAlt(data, PoolSize, formula, alpha = 0.05, verbose = F)
}
\arguments{
\item{data}{A \code{data.frame} with one row for each pooled sampled and columns for
the size of the pool (i.e. the number of specimens / isolates / insects
pooled to make that particular pool), the result of the test of the pool
and any number of columns to be used as the dependent variables in the
logistic regression}

\item{PoolSize}{A \code{string} giving the name of the column with number of specimens/isolates/insects in each pool}

\item{formula}{A \code{formula} of the kind used to define models in \code{glm}. The left-hand
side should be the name of column with the result of each test on each pooled
sample. The result must be stored with 1 indicating a positive test result and
0 indicating a negative test result.}

\item{alpha}{The confidence level to be used for the confidence and credible intervals.
Defaults to 0.5\% (i.e. 95\% intervals)}

\item{verbose}{Logical indicating whether to print progress to screen. Defaults to false
(no printing to screen)}
}
\value{
TBC
}
\description{
Logistic regression with presence/absence tests on pooled samples
}
