% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolReg.R
\name{PoolReg}
\alias{PoolReg}
\title{Frequentist Mixed or Fixed Effect Logistic Regression with Presence/Absence Tests on
Pooled Samples}
\usage{
PoolReg(data, formula, poolSize)
}
\arguments{
\item{data}{A \code{data.frame} with one row for each pooled sampled and
columns for the size of the pool (i.e. the number of specimens / isolates /
insects pooled to make that particular pool), the result of the test of the
pool and any number of columns to be used as the dependent variables in the
logistic regression}

\item{formula}{A \code{formula} of the kind used to define models in
\code{lme4}, which are generalisation of the formulae used in \code{lm} or
\code{glm} that allow for random/group effects. The left-hand side of the
formula should be the name of column in \code{data} with the result of the
test on the pooled samples. The result must be stored with 1 indicating a
positive test result and 0 indicating a negative test result.}

\item{poolSize}{The name of the column with number of
specimens/isolates/insects in each pool}
}
\value{
An object of class \code{glmerMod} (or \code{glm} if there are no
  random/group effects)
}
\description{
It can be useful to do mixed effects logistic regression on the
presence/absence results from pooled samples, however one must adjust for the
size of each pool to correctly identify trends and associations. This can
done using \code{glmer} from the \code{lme4} package and the custom link
function \code{\link{PoolLink}}, defined in this package.
\code{PoolLogitRegMixed} is provided as a more convenient way to fit the same
model. \code{PoolLogitReg} is similar but for the case with only
population/fixed effects. \code{PoolLogitRegMixed} should be used whenever
there is some heirarchical structure to te sampling frame, or in other cases
where a model with group/random effects is appropriate.
}
