% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTestR.R
\name{PoolTestR}
\alias{PoolTestR}
\title{Estimation of prevalence based on results of tests on pooled samples}
\usage{
PoolTestR(data, TestResult, PoolSize, ...)
}
\arguments{
\item{TestResult}{The name of column with the result of each test on each pooled sample}

\item{PoolSize}{The name of the column with number of bugs in each pool}

\item{...}{Optional name(s) of columns with variables to group the data by. If ommitted uses the complete dataset of pooled sample results to calculate a single prevalence. If included a seperate prevalence is calculated for each group defined by these columns}
}
\value{
A tibble
}
\description{
Estimation of prevalence based on results of tests on pooled samples
}
