% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalence.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Predicting Prevalence from a Mixed or Fixed Effect Logistic
Regression with Presence/Absence Tests on Pooled Samples}
\usage{
getPrevalence(model, data = NULL)
}
\arguments{
\item{model}{An object returned by [PoolReg()] or [PoolRegBayes()]}

\item{data}{The data for which prevalence needs to be estimated/predicted. If
not provided, defaults to using the data used to train the model (i.e.
returns the fitted values of the prevalence)}
}
\value{
A \code{list} with at least one field \code{PopulationEffects} and an
  additional field for every random/group effect variable. The field
  \code{PopulationEffects} contains a \code{data.frame} with the prevalence
  estimated based only the fixed/population effects. When the intercept is
  the only fixed/popultion effect, this is just the population mean (possibly
  adjusted for random/group effects). When there are group effects
}
\description{
This function works somewhat like a \code{predict} or \code{fitted} generic
function returning the model predicted prevalence for a given set of data;
however, as the quantity of interest (prevalence) is neither on the response
or link scale we do not use either of these generic functions. Further, when
the model accounts for the heirarchical structure of the sampling frame (e.g.
Region/Village/Site), it is common to want to know the predicted values at
each level of sampling (e.g. Prevalence at each region, village or site) so
these are calculated automatically.
}
\examples{
# Perform logistic-type regression modelling for a synthetic dataset consisting
# of pools (sizes 1, 5, or 10) taken from 4 different regions and 3 different
# years. Within each region specimens are collected at 4 different villages,
# and within each village specimens are collected at 8 different sites.


### Models in a frequentist framework
#ignoring hierarchical sampling frame within each region
Mod <- PoolReg(Result ~ Region + Year,
               data = ExamplePoolData,
               poolSize = NumInPool)
summary(Mod)

#accounting hierarchical sampling frame within each region
HierMod <- PoolReg(Result ~ Region + Year + (1|Village/Site),
                   data = ExamplePoolData,
                   poolSize = NumInPool)
summary(HierMod)
#Extract fitted prevalence for each combination of region and year and then at
#each level of the heirarchical sampling frame (i.e. for each village in each
#region and  each site in each village)
getPrevalence(HierMod)


### Models in a Bayesian framework with default (non-informative) priors
#ignoring hierarchical sampling frame within each region
\dontrun{
BayesMod <- PoolRegBayes(Result ~ Region + Year,
                         data = ExamplePoolData,
                         poolSize = NumInPool)
summary(BayesMod)
getPrevalence(BayesMod) #Extract fitted prevalence for each combination of region and year

#accounting hierarchical sampling frame within each region
BayesHierMod <- PoolRegBayes(Result ~ Region + Year + (1|Village/Site),
                             data = ExamplePoolData,
                             poolSize = NumInPool)
summary(BayesHierMod)
getPrevalence(BayesHierMod)
}

### Calculate adjusted estimates of prevalence
# We use the same function for all four models, but the outputs are slightly different

# Extract fitted prevalence for each combination of region and year
getPrevalence(Mod)
\dontrun{
getPrevalence(BayesMod)
}

#Extract fitted prevalence for each combination of region and year and then at
#each level of the heirarchical sampling frame (i.e. for each village in each
#region and  each site in each village)
getPrevalence(HierMod)
\dontrun{
getPrevalence(BayesHierMod)
}

# You can also use getPrevalence to predict at prevalence for other values of
# the covariates (e.g. predict prevalence in year 4)

#Making a data frame containing data make predict on
ExamplePoolDataFuture <- unique(data.frame(Region = ExamplePoolData$Region,
                                           Village = ExamplePoolData$Village,
                                           Site = ExamplePoolData$Site,
                                           Year = 4))

getPrevalence(Mod, data = ExamplePoolDataFuture)
getPrevalence(HierMod, data = ExamplePoolDataFuture)
\dontrun{
getPrevalence(BayesMod, data = ExamplePoolDataFuture)
getPrevalence(BayesHierMod, data = ExamplePoolDataFuture)
}
}
\seealso{
[PoolReg()] and [PoolRegBayes()]
}
