% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTestR.R
\name{PoolPrev}
\alias{PoolPrev}
\title{Estimation of prevalence based on results of tests on pooled samples}
\usage{
PoolPrev(data, TestResult, PoolSize, ..., alpha = 0.05)
}
\arguments{
\item{TestResult}{The name of column with the result of each test on each pooled sample}

\item{PoolSize}{The name of the column with number of bugs in each pool}

\item{...}{Optional name(s) of columns with variables to group the data by. If ommitted uses the complete dataset of pooled sample results to calculate a single prevalence. If included prevalence is estimated spearately for each group defined by these columns}

\item{alpha}{The confidence level to be used for the confidence and credible intervals. Defaults to 0.5\% (i.e. 95\% intervals)}
}
\value{
A dataframe with columns:
\itemize{
 \item{MLE (the Maximum Likleihood Estimate of prevelance)}
 \item{LR-CI Lower and LR-CI Upper (Lower and Upper Confidence intervals using the Likelihood Ratio method)}
 \item{Bayesian Posterior Expectation}
 \item{Bayesian CI Lower and Bayesian CI Upper}
 \item{Number of Pools}
 \item{Number Positive}
}
When there are no grouping variables (supplied in ...) then the dataframe has only one row with the prevalence estimates for the whole dataset.
When grouping varaibles are supplied, then there is a seperate row for each group.
}
\description{
Estimation of prevalence based on results of tests on pooled samples
}
