% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolRegBayes.R
\name{PoolRegBayes}
\alias{PoolRegBayes}
\title{Bayesian Mixed or Fixed Effect Logistic Regression with Presence/Absence
Tests on Pooled Samples}
\usage{
PoolRegBayes(
  formula,
  data,
  poolSize,
  link = "logit",
  prior = NULL,
  cores = NULL,
  ...
)
}
\arguments{
\item{formula}{A \code{formula} of the kind used to define models in
\code{brms}, which are generalisation of the formulae used in \code{lm,
glm} or \code{lme4}. The left-hand side of the formula should be the name
of column in \code{data} with the result of the test on the pooled samples.
The result must be stored with 1 indicating a positive test result and 0
indicating a negative test result.}

\item{data}{A \code{data.frame} with one row for each pooled sampled and
columns for the size of the pool (i.e. the number of specimens / isolates /
insects pooled to make that particular pool), the result of the test of the
pool and any number of columns to be used as the dependent variables in the
logistic regression.}

\item{poolSize}{The name of the column with number of specimens / isolates /
insects in each pool.}

\item{link}{Link function. There are three options `logit` (i.e logistic
regression, the default), `cloglog` (complementary log-log), and
`loglogit`. The final option blends a log link function and the logit
function so that parameters are (log) prevalence/rate ratios as long as
predicted prevalence is <0.8 (for details see Clark and Barr, Stat Methods
Med Res (2018) <DOI:10.1177/0962280217698174>)}

\item{prior}{The priors to be used for the regression parameters. Defaults to
a non-informative (normal(0,100)) prior on linear coefficients and a
zero-truncated student-t prior on the group effect standard deviations.
Custom priors must be \code{brmsprior} objects produced by
\code{\link[brms:set_prior]{brms::set_prior}}}

\item{cores}{The number of CPU cores to be used. By default one core is used}

\item{...}{Additional arguments to be passed to \code{brms::brms}.}
}
\value{
An object of class \code{brms} with the regression outputs.
}
\description{
It can be useful to do mixed effects logistic regression on the
presence/absence results from pooled samples, however one must adjust for the
size of each pool to correctly identify trends and associations.
}
\examples{
# Perform logistic-type regression modelling for a synthetic dataset consisting
# of pools (sizes 1, 5, or 10) taken from 4 different regions and 3 different
# years. Within each region specimens are collected at 4 different villages,
# and within each village specimens are collected at 8 different sites.


### Models in a frequentist framework
#ignoring hierarchical sampling frame within each region
Mod <- PoolReg(Result ~ Region + Year,
               data = SimpleExampleData,
               poolSize = NumInPool)
summary(Mod)

#accounting hierarchical sampling frame within each region
HierMod <- PoolReg(Result ~ Region + Year + (1|Village) + (1|Site),
                   data = SimpleExampleData,
                   poolSize = NumInPool)
summary(HierMod)


### Models in a Bayesian framework with default (non-informative) priors
#ignoring hierarchical sampling frame within each region
\donttest{
  BayesMod <- PoolRegBayes(Result ~ Region + Year,
                           data = SimpleExampleData,
                           poolSize = NumInPool)
  summary(BayesMod)

  #we could also account for hierarchical sampling frame within each region but
  #note that this is more complex and slower)

  # BayesHierMod <- PoolRegBayes(Result ~ Region + Year + (1|Village) + (1|Site),
  #                              data = SimpleExampleData,
  #                              poolSize = NumInPool)
}

### Calculate adjusted estimates of prevalence
# We use the same function for all four models, but the outputs are slightly different

#For models without hierarchical sampling structure there is an estimate of
#prevalence for every combination of population (fixed) effects: e.g. Region and
#Year
getPrevalence(Mod) #Frequentist model
\donttest{
  getPrevalence(BayesMod) #Bayesian model
}

#For models without hierarchical sampling structure, there is a prevalence
#estimate for each combination of region and year and then at each level of the
#hierarchical sampling frame (i.e. for each village in each region and each site
#in each village)
getPrevalence(HierMod)

# You can also use getPrevalence to predict prevalence for other values of the
# covariates (e.g. predict prevalence in year 4 based on linear trend on the
# logit scale)

#Making a data frame containing data make predictions on
DataFuture <- unique(data.frame(Region = SimpleExampleData$Region,
                                Village = SimpleExampleData$Village,
                                Site = SimpleExampleData$Site,
                                Year = 4))

getPrevalence(Mod, newdata = DataFuture)
getPrevalence(HierMod, newdata = DataFuture)
}
\references{
Clark RG, Barr M: A blended link approach to relative risk regression.
  Statistical Methods in Medical Research 2018, 27(11):3325-3339.
  <DOI:10.1177/0962280217698174>

  Angus McLure, Ben O'Neill, Helen Mayfield, Colleen Lau, Brady McPherson (2021).
  PoolTestR: An R package for estimating prevalence and regression modelling
  for molecular xenomonitoring and other applications with pooled samples.
  Environmental Modelling & Software, 145:105158. <DOI:10.1016/j.envsoft.2021.105158>
}
\seealso{
\code{\link{PoolReg}},
     \code{\link{getPrevalence}}
}
