% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTestR-package.R
\docType{package}
\name{PoolTestR-package}
\alias{PoolTestR-package}
\alias{PoolTestR}
\title{The 'PoolTestR' package.}
\description{
Tools for working with pooled samples. Currently in early stages. Inspired by PoolScreen. Currently only has basic functionally reproducing PoolScreen functionality for R, however more features are planned or in the works: e.g. adjustments for imperfect test specificity/sensitivity; functions for helping with optimal experimental design; functions for inferring whether a disease has been locally eliminated from a series of pooled tests over time
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
