% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{prepareClusterData}
\alias{prepareClusterData}
\title{Prepare hierarchical/clustered survey data for analysis}
\usage{
prepareClusterData(data, result, poolSize, ...)
}
\arguments{
\item{data}{A \code{data.frame} with one row for each pooled sampled and
columns for the size of the pool (i.e. the number of specimens / isolates /
insects pooled to make that particular pool), the result of the test of the
pool. It may also contain additional columns with additional information
(e.g. location where pool was taken) which can optionally be used for
stratifying the data into smaller groups and calculating prevalence by
group (e.g. calculating prevalence for each location)}

\item{result}{The name of column with the result of each test on each pooled
sample. The result must be stored with 1 indicating a positive test result
and 0 indicating a negative test result.}

\item{poolSize}{The name of the column with number of
specimens/isolates/insects in each pool}

\item{...}{Optional name(s) of columns included in the hierarchical sampling
scheme, listed from largest to smallest.}
}
\value{
An object of class \code{data.frame}, containing the same columns as
the input, with a single new column containing a unique identifier for
each location in the survey.

This function checks that each site in the data has a unique identifying
variable.

For example, take a hypothetical survey with two villages \code{A} and 
\code{B}, and two households are sampled within each village. In this survey, 
the \code{"household"} variable is nested inside the \code{"village"} 
variable. Each of the four households must have a unique identifier. 

A common mistake is to reuse names across different locations, e.g., to
name the two households in Village A \code{H1} and \code{H2}, and to name
the two households in Village B \code{H1} and \code{H2}. The household names
are duplicated - there are two houses named \code{H1} and two named 
\code{H2}. To ensure unique variables for each location, the houses could be 
numbered sequentially e.g., \code{c("H1", "H2", "H3", "H4")}, or each house
could be labelled with the corresponding village e.g., 
\code{c("A-H1", "A-H2", "B-H1", "B-H2")}.

The tools in PoolTools apply for all cases where a hierarchical sampling 
frame is involved. The functions do not make assumptions about the 
number of levels present or the names of hierarchical columns. 

Note that each hierarchical variable must have its own column. If the 
sampling scheme is "District" > "Subdistrict" > "Street" > "Unit", there 
should be one column in the data for each of the columns \code{"District"},
\code{"Subdistrict"}, \code{"Street"}, and \code{"Unit"}. 

For the \code{SimpleExampleData} data included in this package, the 
hierarchical sampling scheme is \code{Region} > \code{Village} > \code{Site}.
The function call would be:
\code{prepareClusterData(SimpleExampleData, "Result", "NumInPool",
"Region", "Village", "Site")}
}
\description{
Helper function to prepare data for analysis with \code{HierPoolPrev()} by
ensuring that every location has a unique identifier.
}
\seealso{
\code{\link{HierPoolPrev}}, \code{\link{getPrevalence}}
}
