% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{PrepareClusterData}
\alias{PrepareClusterData}
\title{Prepare hierarchical/clustered survey data for analysis}
\usage{
PrepareClusterData(data, result, poolSize, hierarchy)
}
\arguments{
\item{data}{A \code{data.frame} with one row for each pooled sampled and
columns for the size of the pool (i.e. the number of specimens / isolates /
insects pooled to make that particular pool), the result of the test of the
pool. It may also contain additional columns with additional information
(e.g. location where pool was taken) which can optionally be used for
stratifying the data into smaller groups and calculating prevalence by
group (e.g. calculating prevalence for each location)}

\item{result}{The name of column with the result of each test on each pooled
sample. The result must be stored with 1 indicating a positive test result
and 0 indicating a negative test result.}

\item{poolSize}{The name of the column with number of
specimens/isolates/insects in each pool}

\item{hierarchy}{The name of column(s) indicating the group membership, 
ordered from largest to smallest.}
}
\value{
An object of class \code{data.frame}, containing the same columns as
the input, with a single new column \code{PoolTestR_ID} containing a unique 
identifier for each location in the survey.

In a nested sampling design with multiple levels of grouping, the 
lower-level groups must have names/numbers that differentiate them from all 
other groups at the same level. E.g. If sampling was performed at 200 sites 
across 10 villages (20 site per village), then there should be 200 unique 
names for the sites. If, for instance, the sites are instead numbered 1 to 
20 within each village, the village identifier (e.g. A, B, C...) should be 
combined with the site number to create unique identifiers for each site 
(e.g. A-1, A-2... for sites in village A and B-1, B-2... for the sites in 
village B etc.)

The functions in PoolTools apply for all cases where a hierarchical sampling 
frame is involved, and do not make assumptions about the number of levels 
present or the names of hierarchical columns. 

For the \code{SimpleExampleData} data included in this package, the 
hierarchical sampling scheme is \code{Region} > \code{Village} > \code{Site}.
The function call would be:
\code{prepareClusterData(SimpleExampleData, "Result", "NumInPool",
c("Region", "Village", "Site"))}
}
\description{
Helper function to prepare data for analysis with \code{HierPoolPrev()} by
ensuring that every location has a unique identifier.
}
\seealso{
\code{\link{HierPoolPrev}}, \code{\link{getPrevalence}}
}
